- Se agregan metodos para actualizar tablas y se comienza a trabajar con auth

+++ src/productos/entities/producto.entity.ts
@@ -14,6 +14,6 @@ export class Producto {
   @Column("text")
   categoria: string;
 
-  @Column("float")
+  @Column("int")
   precio: number;
 }
